+++ lib/common/io_bridge_object_store/src/pipeline/buffer.rs
@@ -35,14 +35,45 @@ pub(crate) fn read_into_byte_buffer<A: AsyncRead>(
 
 /// Like [`read_into_byte_buffer`], but fetches the whole object in one GET,
 /// sizing the buffer from the response length (no separate `len`/HEAD).
-pub(crate) fn read_whole_into_byte_buffer<A: AsyncRead>(
+pub(crate) fn read_whole_into_byte_buffer<A: AsyncRead + Clone>(
     file: &BlobFile<A>,
     align: usize,
 ) -> impl Future<Output = Result<AVec<u8, RuntimeAlign>>> + Send + 'static {
-    let whole_fut = file.inner.read_whole(&file.path);
+    read_from_into_byte_buffer(file, 0, align)
+}
+
+/// Like [`read_into_byte_buffer`], but fetches everything from byte offset
+/// `from` to the end of the object in one open-ended GET, sizing the buffer
+/// from the object's total length carried in the response — no separate
+/// `len`/HEAD round-trip on the happy path. `from == 0` reads the whole object.
+///
+/// An offset at or past the end has no tail to read. The backend reports that as
+/// an unsatisfiable-range error rather than an empty body, so the error path
+/// confirms with a single `len`: if `from >= eof` the tail is genuinely empty
+/// and we yield a zero-length buffer; otherwise the original read error stands.
+pub(crate) fn read_from_into_byte_buffer<A: AsyncRead + Clone>(
+    file: &BlobFile<A>,
+    from: u64,
+    align: usize,
+) -> impl Future<Output = Result<AVec<u8, RuntimeAlign>>> + Send + 'static {
+    let read_fut = file.inner.read_from(&file.path, from);
+    // Cloned for the cold disambiguation path only; building the `len` future is
+    // deferred until a read error actually occurs.
+    let inner = file.inner.clone();
+    let path = file.path.clone();
     async move {
-        let (size, stream) = whole_fut.await?;
-        let buf = avec_rt!([align] | 0u8; size as usize);
+        let (size, stream) = match read_fut.await {
+            Ok(ok) => ok,
+            Err(err) => {
+                let eof = inner.len(&path).await?;
+                if from >= eof {
+                    return Ok(avec_rt!([align] | 0u8; 0));
+                }
+                return Err(err);
+            }
+        };
+        let len = size.saturating_sub(from) as usize;
+        let buf = avec_rt!([align] | 0u8; len);
         fold_stream_into_buffer(stream, buf).await
     }
 }
