+++ lib/common/io_bridge_object_store/src/source.rs
@@ -166,15 +166,24 @@ impl<S: BlobBackend> AsyncRead for Arc<S> {
         }
     }
 
-    fn read_whole(
+    fn read_from(
         &self,
         path: &Path,
+        from: u64,
     ) -> impl Future<Output = Result<(u64, BoxStream<'static, Result<Bytes>>)>> + Send + 'static
     {
         let store = self.clone();
         let key = build_key(path);
         async move {
-            let result = store.get(&key).await.map_err(|err| match err {
+            // `from == 0` is a plain whole-object GET; a positive offset asks the
+            // backend for everything from `from` onward in a single open-ended
+            // GET (`Range: bytes=from-`). Either way the response carries the
+            // object's total size, so no separate HEAD is needed.
+            let opts = GetOptions {
+                range: (from > 0).then_some(GetRange::Offset(from)),
+                ..Default::default()
+            };
+            let result = store.get_opts(&key, opts).await.map_err(|err| match err {
                 object_store::Error::NotFound { .. } => UniversalIoError::NotFound {
                     path: PathBuf::from(key.to_string()),
                 },
